service/processing-service/requirements.txt
# Core
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pika==1.3.2
minio>=7.1.0
certifi>=2024.4.0
python-dotenv>=1.0.0
pydantic>=2.1.1

# Image processing (headless for container)
opencv-python-headless>=4.8.0.74
numpy>=1.26.0

# Optional future additions
prometheus-client>=0.17.0  # for metrics if added later
python-json-logger>=2.0.7